requirements.txt
appdirs==1.4.3
attrs==18.2.0
black==18.9b0
Click==7.0
flake8==3.6.0
mccabe==0.6.1
Pillow==5.4.1
pkg-resources==0.0.0
pycodestyle==2.4.0
pyflakes==2.0.0
toml==0.10.0
